+++ src/libs/conduit/conduit_data_array.hpp
@@ -103,6 +103,14 @@ public:
                                     Node &info,
                                     const float64 epsilon = CONDUIT_EPSILON) const;
 
+    ///
+    /// Summary Stats Helpers
+    ///
+    T               min()  const;
+    T               max()  const;
+    T               sum() const;
+    float64         mean() const;
+
 //-----------------------------------------------------------------------------
 // Setters
 //-----------------------------------------------------------------------------
@@ -207,7 +215,7 @@ public:
     #endif
 
     #ifndef CONDUIT_USE_DOUBLE
-       void setconst std::initializer_list<double> &values);
+       void set(const std::initializer_list<double> &values);
     #endif
 
     //-------------------------------------------------------------------------
@@ -311,7 +319,14 @@ public:
     void            to_yaml_stream(std::ostream &os) const;
 
     void            compact_elements_to(uint8 *data) const;
-    
+
+    /// Creates a string repression for printing that limits
+    /// the number of elements shown to a max number
+    std::string     to_summary_string_default() const;
+    std::string     to_summary_string(index_t threshold=5) const;
+    void            to_summary_string_stream(std::ostream &os,
+                                             index_t threshold=5) const;
+
 //-----------------------------------------------------------------------------
 // -- stdout print methods ---
 //-----------------------------------------------------------------------------