+++ src/tests/conduit/python/t_python_conduit_node.py
@@ -535,6 +535,49 @@ class Test_Conduit_Node(unittest.TestCase):
         n['vs'].set_external(v[:,0,0])
         n['vs_expected'] = np.array(v[:,0,0],np.float64)
 
+
+    def test_describe(self):
+        n = Node()
+        n["a"] = [1,2,3,4,5];
+        n["b"] = [1,2,3];
+        n["c"] = [1,2,3,4,5,6];
+        n["d"] = [1,2,3,4,5,6,7];
+        n["e"] = [1,2,3,4,5,6,7,8,9,10,11,12];
+        n["f"] = [1.0,2.0,3.0,4.0,5.0,6.0,7.0];
+        n["g"] = [2.0,4.0];
+
+        d = n.describe()
+        print(d)
+
+        self.assertEqual(d["a/count"],5);
+        self.assertEqual(d["b/count"],3);
+        self.assertEqual(d["c/count"],6);
+        self.assertEqual(d["d/count"],7);
+        self.assertEqual(d["e/count"],12);
+        self.assertEqual(d["f/count"],7);
+
+        self.assertEqual(d["a/min"],1)
+        self.assertEqual(d["b/min"],1)
+        self.assertEqual(d["c/min"],1)
+        self.assertEqual(d["d/min"],1)
+        self.assertEqual(d["e/min"],1)
+        self.assertEqual(d["f/min"],1.0)
+
+        self.assertEqual(d["a/max"],5)
+        self.assertEqual(d["b/max"],3)
+        self.assertEqual(d["c/max"],6)
+        self.assertEqual(d["d/max"],7)
+        self.assertEqual(d["e/max"],12)
+        self.assertEqual(d["f/max"],7.0)
+
+        self.assertEqual(d["g/mean"],3.0);
+
+        opts = Node()
+        opts["threshold"] = 10
+        d = n.describe(opts)
+        print(d)
+
+
 if __name__ == '__main__':
     unittest.main()